backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0

# LangChain ecosystem
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
langchain-anthropic==0.1.1

# Vector database
chromadb==0.4.22

# Document processing
pypdf==3.17.4
python-docx==1.1.0
python-magic==0.4.27
beautifulsoup4==4.12.3
requests==2.31.0

# Embeddings
sentence-transformers==2.3.1
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
